move HttpServer to a separate file
